radio_mac.c: Fix two compiler warnings about an overflow in an implicit integer conversion.

+++ libraries/src/radio_mac/radio_mac.c
@@ -142,7 +142,7 @@ ISR(RF, 0)
         // We were not reading data from the radio fast enough, so there was
         // a RX overflow.  This should not happen.  Report it as an error.
         radioRxOverflowOccurred = 1;
-        RFIF = ~0x40;
+        RFIF = (uint8)(~0x40);
 
         // The radio module is probably now in the RX_OVERFLOW state where it can not
         // receive packets.  The way to get out of this state is:
@@ -179,7 +179,7 @@ void radioMacEvent(uint8 event)
     // We want to do it before restarting the radio (to avoid accidentally missing
     // an event) but we want to do it as long as possible AFTER turning off the
     // radio.
-    RFIF = ~0x30;  // Clear IRQ_DONE and IRQ_TIMEOUT if they are set.
+    RFIF = (uint8)(~0x30);  // Clear IRQ_DONE and IRQ_TIMEOUT if they are set.
 
     /** Start up the radio in the new state which was decided above. **/
     switch(radioMacState)
